feat: make buildとmake runをdockerの外でも実行できるようにする
別にbuildとrun自体はdockerの外で実行されてもいい気がする

+++ scripts/dev.sh
@@ -1,3 +1,7 @@
 #!/usr/bin/env bash
 
-sudo docker compose run --build --rm -it dev
+if [ -f "/.dockerenv" ]; then
+    echo "cannot launch docker inside docker" 2>&1
+else
+    sudo docker compose run --build --rm -it dev
+fi
